requirements.txt
google-genai
pandas
psycopg2-binary
rich
sqlalchemy
python-dotenv